Create different refreshRequestDebouncer instance for different recyclerlistview instances
Incase of multiple recyclerlistviews when we were using same instance of refreshRequestDebouncer,
one of the instance was using the dimensions(state) of other instance.

To Reproduce the issue:
Add two instances of recycler list view one below the other with couple of elements.
The spacing between the elements was getting miscalculated.